overhaul setupUnit to setupUnitMocks.ts, remove useMock in favor of global setup
