+++ test/BasemapSession.test.ts
@@ -1,9 +1,8 @@
 //@ts-nocheck
-import { describe, expect, vi, beforeAll, beforeEach } from 'vitest';
+import { describe, expect, vi } from 'vitest';
 import { customTest } from './BaseTest'
 import { BasemapSession } from '../src/BasemapSession';
 import { BasemapStyleSession } from '@esri/arcgis-rest-basemap-sessions';
-import { useMock, removeMock } from './setupUnit';
 import sessionResponseRaw from './mock/BasemapSession/valid-session.json';
 import expiredSessionResponseRaw from './mock/BasemapSession/expired-session.json';
 import { ApiKeyManager } from '@esri/arcgis-rest-request';
@@ -33,16 +32,6 @@ const test = customTest.extend({
   }
 })
 describe('Basemap session unit tests', () => {
-  beforeAll(async () => {
-    useMock();
-    return () => {
-      removeMock();
-    }
-  });
-  beforeEach(() => {
-    fetchMock.resetMocks();
-  });
-
   test('Uses ArcGIS REST JS to start a basemap session.', async ({apiKey}) => {
     fetchMock.once(sessionResponse);
